refactor(editor): extract styles, entity option helpers and host contract
Mechanical module split, step 1/4: editor-styles.ts (verbatim CSS move),
entity-options.ts (pure functions, this._hass becomes a parameter),
editor-host.ts (StrategyEditorHost interface the panel modules will
program against). No behavior change.
